[sarldoc] Add dynamic class loader.
