+++ products/sarldoc/src/main/java/io/sarl/sarldoc/modules/commands/Messages.java
@@ -34,11 +34,14 @@ import org.eclipse.osgi.util.NLS;
 @SuppressWarnings("all")
 public class Messages extends NLS {
 	private static final String BUNDLE_NAME = Messages.class.getPackage().getName() + ".messages"; //$NON-NLS-1$
-	public static String SarldocCommandModuleProvider_0;
+
 	static {
 		// initialize resource bundle
 		NLS.initializeMessages(BUNDLE_NAME, Messages.class);
 	}
+
+	public static String SarldocCommandModuleProvider_0;
+	public static String SarldocFakeCommandModuleProvider_0;
 
 	private Messages() {
 	}
