+++ products/sarldoc/src/main/java/io/sarl/sarldoc/modules/internal/SarldocDynamicClassLoaderModuleProvider.java
@@ -21,30 +21,33 @@
 
 package io.sarl.sarldoc.modules.internal;
 
-import org.eclipse.osgi.util.NLS;
+import com.google.inject.Module;
+import io.bootique.BQModule;
+import io.bootique.BQModuleProvider;
 
-/** Messages for the SARL batch compiler.
+/** Provider of the module for injecting the dynamic class loader.
  *
  * @author $Author: sgalland$
  * @version $FullVersion$
  * @mavengroupid $GroupId$
  * @mavenartifactid $ArtifactId$
- * @ExcludeFromApidoc
+ * @since 0.10
  */
-@SuppressWarnings("all")
-public class Messages extends NLS {
-	private static final String BUNDLE_NAME = Messages.class.getPackage().getName() + ".messages"; //$NON-NLS-1$
-	public static String SarldocDefaultCommandModuleProvider_0;
-	public static String SarldocApplicationModule_0;
-	public static String SarldocApplicationModule_1;
-	public static String SarldocApplicationModule_2;
-	public static String SarldocApplicationModuleProvider_0;
-	public static String SarldocDynamicClassLoaderModuleProvider_0;
-	static {
-		// initialize resource bundle
-		NLS.initializeMessages(BUNDLE_NAME, Messages.class);
+public class SarldocDynamicClassLoaderModuleProvider implements BQModuleProvider {
+
+	@Override
+	public Module module() {
+		return new SarldocDynamicClassLoaderModule();
 	}
 
-	private Messages() {
-	}
+	@Override
+    public BQModule.Builder moduleBuilder() {
+        return BQModule
+                .builder(module())
+                .overrides(overrides())
+                .providerName(name())
+                .configs(configs())
+                .description(Messages.SarldocDynamicClassLoaderModuleProvider_0);
+    }
+
 }
