+++ products/sarldoc/src/main/resources/META-INF/services/io.bootique.BQModuleProvider
@@ -1,2 +1,4 @@
+io.sarl.sarldoc.modules.internal.SarldocDynamicClassLoaderModuleProvider
 io.sarl.sarldoc.modules.configs.SarldocConfigModuleProvider
+io.sarl.sarldoc.modules.commands.SarldocFakeCommandModuleProvider
 io.sarl.sarldoc.modules.commands.SarldocCommandModuleProvider
