+++ products/sarldoc/src/main/resources/io/sarl/sarldoc/configs/messages.properties
@@ -9,3 +9,4 @@ Placement_7 =  into the fields'' documentations
 SarldocConfig_0 = The environment variable JAVA_HOME is not correctly set.
 SarldocConfig_1 = The environment variable JAVA_HOME={0} doesn't exist or is not a valid directory.
 SarldocConfig_2 = The javadoc executable ''{0}'' doesn't exist or is not a file. Verify the JAVA_HOME environment variable.
+SarldocConfig_3 = The mandatory library ''tools.jar'' doesn't exist or is not a file. Verify your installation of your Java Development Kit (JDK).
