+++ products/sarldoc/src/main/resources/io/sarl/sarldoc/modules/commands/messages.properties
@@ -1,1 +1,2 @@
 SarldocCommandModuleProvider_0 = The command for starting sarldoc.
+SarldocFakeCommandModuleProvider_0 = The fake command for starting sarldoc.
