+++ products/sarldoc/src/main/resources/io/sarl/sarldoc/modules/internal/messages.properties
@@ -10,3 +10,4 @@ SarldocApplicationModule_2 = Sarldoc is a documentation generator for the SARL l
 	a separate destination directory with the {3} option.\nBy default, the generated HTML pages in the directory {4}. You can specify \
 	a separate destination directory with the {5} option.
 SarldocApplicationModuleProvider_0 = The name and description of the application.
+SarldocDynamicClassLoaderModuleProvider_0 = The dynamic class loader for sarldoc.
